+++ packages/core/src/runtime/index.ts
@@ -9,7 +9,7 @@ import {
   normalizeRpcResponse,
 } from '../protocol/index.js'
 
-export const CODY_WEB_CORE_VERSION = '0.37.10'
+export const CODY_WEB_CORE_VERSION = '0.37.12'
 
 export type { RuntimeNotification, ServerRequest } from '../protocol/index.js'